+++ cpp/src/community/ktruss.cu
@@ -35,8 +35,9 @@ namespace cugraph {
 namespace detail {
 
 template <typename VT, typename ET, typename WT>
-std::unique_ptr<experimental::GraphCOO<VT, ET, WT>> ktruss_subgraph_impl(
-  experimental::GraphCOOView<VT, ET, WT> const &graph, int k, rmm::mr::device_memory_resource *mr)
+std::unique_ptr<GraphCOO<VT, ET, WT>> ktruss_subgraph_impl(GraphCOOView<VT, ET, WT> const &graph,
+                                                           int k,
+                                                           rmm::mr::device_memory_resource *mr)
 {
   using HornetGraph = hornet::gpu::Hornet<VT>;
   using UpdatePtr   = hornet::BatchUpdatePtr<VT, hornet::EMPTY, hornet::DeviceType::DEVICE>;
@@ -67,7 +68,7 @@ std::unique_ptr<experimental::GraphCOO<VT, ET, WT>> ktruss_subgraph_impl(
   kt.runForK(k);
   CUGRAPH_EXPECTS(cudaPeekAtLastError() == cudaSuccess, "KTruss : Failed to run");
 
-  auto out_graph = std::make_unique<experimental::GraphCOO<VT, ET, WT>>(
+  auto out_graph = std::make_unique<GraphCOO<VT, ET, WT>>(
     graph.number_of_vertices, kt.getGraphEdgeCount(), graph.has_data(), stream, mr);
 
   kt.copyGraph(out_graph->src_indices(), out_graph->dst_indices());
@@ -78,8 +79,8 @@ std::unique_ptr<experimental::GraphCOO<VT, ET, WT>> ktruss_subgraph_impl(
   return out_graph;
 }
 template <typename VT, typename ET, typename WT>
-std::unique_ptr<experimental::GraphCOO<VT, ET, WT>> weighted_ktruss_subgraph_impl(
-  experimental::GraphCOOView<VT, ET, WT> const &graph, int k, rmm::mr::device_memory_resource *mr)
+std::unique_ptr<GraphCOO<VT, ET, WT>> weighted_ktruss_subgraph_impl(
+  GraphCOOView<VT, ET, WT> const &graph, int k, rmm::mr::device_memory_resource *mr)
 {
   using HornetGraph = hornet::gpu::Hornet<VT, hornet::EMPTY, hornet::TypeList<WT>>;
   using UpdatePtr   = hornet::BatchUpdatePtr<VT, hornet::TypeList<WT>, hornet::DeviceType::DEVICE>;
@@ -110,7 +111,7 @@ std::unique_ptr<experimental::GraphCOO<VT, ET, WT>> weighted_ktruss_subgraph_imp
   kt.runForK(k);
   CUGRAPH_EXPECTS(cudaPeekAtLastError() == cudaSuccess, "KTruss : Failed to run");
 
-  auto out_graph = std::make_unique<experimental::GraphCOO<VT, ET, WT>>(
+  auto out_graph = std::make_unique<GraphCOO<VT, ET, WT>>(
     graph.number_of_vertices, kt.getGraphEdgeCount(), graph.has_data(), stream, mr);
 
   kt.copyGraph(out_graph->src_indices(), out_graph->dst_indices(), out_graph->edge_data());
@@ -124,8 +125,9 @@ std::unique_ptr<experimental::GraphCOO<VT, ET, WT>> weighted_ktruss_subgraph_imp
 }  // namespace detail
 
 template <typename VT, typename ET, typename WT>
-std::unique_ptr<experimental::GraphCOO<VT, ET, WT>> k_truss_subgraph(
-  experimental::GraphCOOView<VT, ET, WT> const &graph, int k, rmm::mr::device_memory_resource *mr)
+std::unique_ptr<GraphCOO<VT, ET, WT>> k_truss_subgraph(GraphCOOView<VT, ET, WT> const &graph,
+                                                       int k,
+                                                       rmm::mr::device_memory_resource *mr)
 {
   CUGRAPH_EXPECTS(graph.src_indices != nullptr, "Graph source indices cannot be a nullptr");
   CUGRAPH_EXPECTS(graph.dst_indices != nullptr, "Graph destination indices cannot be a nullptr");
@@ -137,14 +139,10 @@ std::unique_ptr<experimental::GraphCOO<VT, ET, WT>> k_truss_subgraph(
   }
 }
 
-template std::unique_ptr<experimental::GraphCOO<int32_t, int32_t, float>>
-k_truss_subgraph<int, int, float>(experimental::GraphCOOView<int, int, float> const &,
-                                  int,
-                                  rmm::mr::device_memory_resource *);
+template std::unique_ptr<GraphCOO<int32_t, int32_t, float>> k_truss_subgraph<int, int, float>(
+  GraphCOOView<int, int, float> const &, int, rmm::mr::device_memory_resource *);
 
-template std::unique_ptr<experimental::GraphCOO<int32_t, int32_t, double>>
-k_truss_subgraph<int, int, double>(experimental::GraphCOOView<int, int, double> const &,
-                                   int,
-                                   rmm::mr::device_memory_resource *);
+template std::unique_ptr<GraphCOO<int32_t, int32_t, double>> k_truss_subgraph<int, int, double>(
+  GraphCOOView<int, int, double> const &, int, rmm::mr::device_memory_resource *);
 
 }  // namespace cugraph