+++ CHANGELOG.md
@@ -27,6 +27,7 @@
 - PR #975 Upgrade GitHub template
 - PR #976 Fix error in Graph.edges(), update cuDF rename() calls
 - PR #977 Update force_atlas2 to call on_train_end after iterating
+- PR #987 Move graph out of experimental namespace
 - PR #984 Removing codecov until we figure out how to interpret failures that block CI
 - PR #985 Add raft handle to BFS, BC and edge BC